Added tab view
